requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1

# LangChain (versões compatíveis entre si)
langchain-community==0.2.10
langchain-core==0.2.43
langchain-text-splitters==0.2.4

# Embeddings (API + ETL usam o mesmo modelo)
sentence-transformers==2.2.2
huggingface-hub==0.24.6
transformers==4.48.3

# Vetores / similaridade
#faiss-cpu==1.7.4
rapidfuzz==3.9.7
numpy==1.26.4
#torch==2.8.0

# PDF
pypdf==4.2.0

#DOCX
python-docx==1.1.2


# Gemini (SDK estável)
google-generativeai>=0.7.2,<0.8

# utilidades
orjson==3.11.3
requests==2.32.5
tenacity==8.5.0
SQLAlchemy==2.0.43
aiohttp==3.12.15
dataclasses-json==0.6.7

# NLTK (seu pipeline usa)
nltk==3.9.1
